+++ module/combatTracker.mjs
@@ -118,9 +118,10 @@ export class HeroSystem6eCombatTracker extends CombatTracker {
         context.segments = [];
         for (let s = 1; s <= 12; s++) {
             context.segments[s] = [];
-            for (let turn of context.turns) {
+            for (let [t, turn] of context.turns.entries()) {
                 if (turn.flags.segment === s) {
                     context.segments[s].push(turn);
+                    turn.flags.turnNumber = t;
                 }
             }
         }