+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Statuses can now be toggled off from combat tracker.
 - Fixed issue where adding several combatants sometimes messes up current phase/segment.
 - Defeated tokens now have a red tint on their image, are slightly transparent, do not show any effects (except for skull overlay) and are sent to the back of the token order. [#2033](https://github.com/dmdorman/hero6e-foundryvtt/issues/2033)
+- Fix for 5e actors uploading as PRONE [#2050](https://github.com/dmdorman/hero6e-foundryvtt/issues/2050)
 
 ## Version 4.0.23 20250315